Simple corrections and added missing movement results
